Harden backend quality gates and close the scripts hole
`backend/scripts/` (3.3k LOC of billing backfills, reconciliation,
repricing and provisioning) was outside mypy, the CC/LOC policy and
Vulture. Widening required fixing the gate itself: `compare_policies`
rejected ANY change to `roots`, so the check meant to prevent weakening
also permanently blocked widening. It now rejects removals only.

That immediately caught two dead seams in the dev seeder, both silent:

* It patched `app.workers.audit_worker.build_adapter`, an attribute that
  stopped existing in 3a629f8c when execution moved to
  `app.workers.audit.execution`. The assignment created a new attribute
  nobody read, so seeded audits called real providers with fake keys.
* It passed `httpx.MockTransport` where `SiteHealthWorker` wants an
  `AcquisitionTransport`, so every seeded crawl fetch would have raised
  `AttributeError` on the missing `fetch`.

Suppression hygiene. RUF100 is enabled, so a `# noqa` that suppresses
nothing now fails the build. The 22 dead directives were three blanket
`# ruff: noqa: E501` on files with no long lines, one stale `B008`, three
for rules never enabled, and fifteen `# noqa: BLE001` for a rule enabled
in one subpackage only. `BLE` is now repo-wide, which made those fifteen
real. Of the 27 unsuppressed blind excepts: six narrowed to real
exception types, six routed through the existing
`analysis/site_health/dom.py` owner (`brand_evidence.py` was
re-implementing the fail-open-without-logging pattern that module exists
to fix), and fifteen given directives that name their reason.

Eight `assert` statements were import-time drift guards between
`Literal[...]` aliases and their config vocabularies. `python -O` strips
those, so the one check keeping an API literal in lock-step with its
persisted catalog vanished in the optimized build. They raise now, via
`app/core/literals.py`.

Eight complexity violations were refactored rather than excepted; the
policy forbids new exceptions and that stays true. `seed_dev_data.py`
went 805 -> 644 lines with its worker drains extracted to
`seed_dev_runs.py`, and `seed()` from CC 29 to an orchestrator.

New gates: Ruff S/ASYNC/BLE/RUF, import-linter (seven contracts, three
warts named rather than ruled away), deptry (which found `pydantic` and
`starlette` imported by ~30 modules but declared nowhere, arriving only
because FastAPI pulled them in), diff-cover on changed lines, and mypy
`check_untyped_defs` + `strict_equality`.

`warn_unreachable` is deliberately off: most of its findings are
defensive guards lxml's stubs declare impossible, and the one that
matters guards a `visibility_score` the ORM maps as non-nullable, so
deleting it would erase the unknown-vs-zero distinction the invariants
require. `warn_return_any` (70 findings) is a separate slice.

Also removed the empty `app/schemas/` package, and linted `reset-db.py`
and `docs/validate_documentation.py` for the first time -- Ruff runs with
`backend/` as cwd, so nothing had ever seen them.
`tests/unit/test_static_analysis_tools.py` now asserts the gate
configuration itself, so weakening one of these breaks a test.

+++ backend/tests/unit/test_billing_quotes.py
@@ -113,7 +113,7 @@ def test_india_quote_applies_configured_gst(monkeypatch) -> None:
     assert (
         quote.total_price.amount_minor
         == quote.base_price.amount_minor + quote.tax.amount_minor
-        == 9_900 * 83 + int(round(9_900 * 83 * 0.18))
+        == 9_900 * 83 + round(9_900 * 83 * 0.18)
     )
